Playwright Create testcases

+++ Backend/Api/Infrastructure/Model/Box.cs
@@ -26,5 +26,5 @@ public class Box
     public double Width { get; set; }
     [Required(ErrorMessage = "Width is required")]
     [Range(0, double.MaxValue, ErrorMessage = "Value must be non-negative")]
-    public decimal Height { get; set; }
+    public double Height { get; set; }
 }
